Merge branch 'main' into arkadiusz-dev

+++ Spring/src/main/java/pl/ttsw/GameRev/repository/GameRepository.java
@@ -6,9 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.ttsw.GameRev.model.Game;
 
-
 @Repository
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
     Game findGameByTitle(String title);
     Game findGameById(Long id);
     boolean existsById(Long id);
